seamcarving/api/requirements.txt
numpy
scipy
matplotlib
Pillow
numba
flask
flask-cors